+++ wsep192/src/FinancialSystemImpl.cs
@@ -15,12 +15,12 @@ namespace src
             return true;
         }
 
-        public bool payment(long cardNumber, DateTime date, int amount, int paymentTarget)
+        public bool payment(long cardNumber, DateTime date, double amount, int paymentTarget)
         {
             return true;
         }
 
-        public bool Chargeback(long cardNumber, DateTime date, int amount)
+        public bool Chargeback(long cardNumber, DateTime date, double amount)
         {
             return true;
         }
